Добавил абстрактную фабрику
