+++ factory/constants.ts
@@ -1,0 +1,9 @@
+enum PizzaTypes {
+    Cheese = 'cheese',
+    Pepperoni = 'pepperoni',
+    Clam = 'clam',
+    Veggie = 'veggie',
+    Greek = 'greek'
+};
+
+export { PizzaTypes }
